feat:user menambahkan alert

+++ resources/views/pages/data-master/user/_functions/function.blade.php
@@ -20,25 +20,46 @@
                 $('#id').val(res.id);
                 $('#nama').val(res.nama);
                 $('#email').val(res.email);
-                $('#password').val(res.password);
+                // $('#password').val(res.password);
                 $('#role').val(res.role);
                 $('#division_id').val(res.division_id);
             }
         });
     }
 
+
     function deleteFunc(id) {
         var id = id;
-        $.ajax({
-            type: "DELETE",
-            url: "{{ route('user.destroy') }}",
-            data: {
-                id: id
-            },
-            dataType: 'json',
-            success: function(res) {
-                var oTable = $('#myTable').dataTable();
-                oTable.fnDraw(false);
+        Swal.fire({
+            title: 'Apakah Anda yakin?',
+            text: "Data yang dihapus tidak dapat dikembalikan!",
+            icon: 'warning',
+            showCancelButton: true,
+            confirmButtonColor: '#d33',
+            cancelButtonColor: '#3085d6',
+            confirmButtonText: 'Ya, Hapus!',
+            cancelButtonText: 'Batal'
+        }).then((result) => {
+            if (result.isConfirmed) {
+                $.ajax({
+                    type: "DELETE",
+                    url: "{{ route('user.destroy') }}",
+                    data: {
+                        id: id
+                    },
+                    dataType: 'json',
+                    success: function(res) {
+                        var oTable = $('#myTable').dataTable();
+                        oTable.fnDraw(false);
+                        Swal.fire({
+                            icon: 'success',
+                            title: 'Success',
+                            text: res.message,
+                            timer: 2000,
+                            showConfirmButton: false
+                        })
+                    }
+                });
             }
         });
     }
@@ -60,6 +81,23 @@
                 oTable.fnDraw(false);
                 $("#btn-save").html('Submit');
                 $("#btn-save").attr("disabled", false);
+                if (response.success) {
+                    Swal.fire({
+                        icon: 'success',
+                        title: 'Success',
+                        text: response.message,
+                        timer: 1000,
+                        showConfirmButton: false
+                    });
+                } else if (!response.success) {
+                    Swal.fire({
+                        icon: 'error',
+                        title: 'Failed',
+                        text: response.message,
+                        timer: 2000,
+                        showConfirmButton: false
+                    });
+                }
             },
             error: function(data) {
                 console.log(data);
